Bug 1417794 - move Adts.h, AnnexB.h and H264.h to dom/media/platform/agnostic/bytestreams. r=kinetik
