+++ dom/media/gmp/ChromiumCDMParent.cpp
@@ -18,8 +18,8 @@
 #include "mozilla/dom/MediaKeyMessageEventBinding.h"
 #include "mozilla/gmp/GMPTypes.h"
 #include "mozilla/Unused.h"
-#include "mp4_demuxer/AnnexB.h"
-#include "mp4_demuxer/H264.h"
+#include "AnnexB.h"
+#include "H264.h"
 
 namespace mozilla {
 namespace gmp {
