+++ dom/media/mp4/DecoderData.cpp
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "mp4_demuxer/Adts.h"
-#include "mp4_demuxer/AnnexB.h"
+#include "Adts.h"
+#include "AnnexB.h"
 #include "mp4_demuxer/BufferReader.h"
 #include "mp4_demuxer/DecoderData.h"
 #include "mozilla/ArrayUtils.h"
