+++ dom/media/platforms/PDMFactory.cpp
@@ -46,7 +46,7 @@
 #include "MP4Decoder.h"
 #include "mozilla/dom/RemoteVideoDecoder.h"
 
-#include "mp4_demuxer/H264.h"
+#include "H264.h"
 
 #include <functional>
