+++ dom/media/platforms/agnostic/gmp/GMPVideoDecoder.cpp
@@ -9,7 +9,7 @@
 #include "GMPVideoHost.h"
 #include "MediaData.h"
 #include "mozilla/EndianUtils.h"
-#include "mp4_demuxer/AnnexB.h"
+#include "AnnexB.h"
 #include "MP4Decoder.h"
 #include "prsystem.h"
 #include "VPXDecoder.h"
