+++ dom/media/platforms/moz.build
@@ -37,9 +37,9 @@ UNIFIED_SOURCES += [
 ]
 
 DIRS += [
+    'agnostic/bytestreams',
     'agnostic/eme',
     'agnostic/gmp',
-    'agnostic/mp4_demuxer',
     'omx'
 ]
